Made some corrections:'
This one was not declared:
> LIST Der/n = Der/n ;

Added TV to the Actio Ess so it will not hit IV:
< SET V-TRANS-ACT-NOT-ACT = V-TRANS - Der/Pass - Actio - Actor - PrsPrc - Attr \
| (Actio Ess) ;
---
> SET V-TRANS-ACT-NOT-ACT = V-TRANS - Der/Pass - Der/n - Actio - Actor - PrsPrc\
 - Attr | (TV Actio Ess) ;

Added Der/n to this set:
< SET V-TRANS-ACT-NOT-ACT-NOT-AUX = V-TRANS - Der/Pass - Actio - Actor - AUX | \
(Actio Ess) ;
---
> SET V-TRANS-ACT-NOT-ACT-NOT-AUX = V-TRANS - Der/Pass - Der/n - Actio - Actor \
- AUX | (Actio Ess) ;

Added "muitalit" to the STVLIST.

Made a new rule:
> SELECT ("gilvu") IF (0 Adv LINK *-1 ("vuoitit") BARRIER SV-BOUNDARY);

Worked with Px:
Removed "siessi" from N-NOPXLIST because of the homonymi with "siessá" which can have Px.

Made a new set:
> SET S-BOUNDARY1  = (Pron Interr) | (Pron Rel) | ("muhto") | MO | ("\;") | (":\
") | ("-") | ("â") ;
>       # includes not CS because of "go" in questions, before it is disambuate\
d.

and changed with S-BOUNDARY in two rules.

Widened this rule:
< REMOVE Px IF (0 N-NOPX);
---
> REMOVE Px IF (0 N-NOPX OR Num OR Der/n);

Made a new rule:
> SELECT (PxSg2) IF (0 (N Pl Nom))(*-1 (V Sg2) BARRIER S-BOUNDARY)(*-1 V-TRANS-\
ACT-NOT-ACT BARRIER SV-BOUNDARY OR Acc LINK NOT 0 A OR Imprt);
>       ## Leat go oaidnán oappát?

Made a rule for a special case:
> REMOVE Loc IF (0 ("ieÅ¡#dovdu") LINK 0 Px) ;
>       ## Son rahčá ieÅ¡dovdduinis.
>