Moved mappingrules so verb- and CC/CS and CS-COMP are after the verbdisambiguation, and the rest is right before casedisambuation. I think this is a good solution.
Tidied the SV-BOUNDARYsets so now is the SV-BOUNDARY without synt.tags, and SV-BOUNDARY2 continues synt.tags for CC and verbs. Removed the SV-BOUNDARY-NO-SYNT.

There is some work to do with the mappingrules. Many of them can now be removed.

Made a new rule for VOC:
> MAP (@VOC) TARGET HUMAN + Nom IF (-1 BOS OR REALCLB)(1 COMMA)(NEGATE 2 HUMAN LINK 1 COMMA OR CC)(NEGATE 2 Rel) ;
and removed the older ones.