THIS IS AN IMPORTANT CHANGE!
I commented out the +UseN/Circ tag from the numerals.
The reason for it was that we remove +UseN/Circ from
the Apertium fst. Now, all numerals (also simplex okta, 
guokte, ...) will be excluded from all automata that
discard +UseN/Circ, and I am not convinced that is what
any of our applications intend. We will test this out
for Apertium, but Divvun folks, please look at this.