THIS IS AN IMPORTANT CHANGE!
I commented out the +UseN/Circ tag from the numerals.
The reason for it was that we remove +UseN/Circ from
the Apertium fst. Now, all numerals (also simplex okta, 
guokte, ...) will be excluded from all automata that
discard +UseN/Circ, and I am not convinced that is what
any of our applications intend. We will test this out
for Apertium, but Divvun folks, please look at this.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@32851 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 112e2e331797beeb7f09b038c0d4bc50efa4df1a