A major change to get rid of the last remnants off M4 processing and prepare for the updated Hunspell speller compilation. Details:
* the bare-bones lexical transducer is now built with hyphens and morphological boundaries visible, that is, the twolc file corresponds to the old hyph-twolc.bin file.
* all non-visible morphological and other boundaries are removed when the lexical transducer is converted to the traditional *.save file
* all speller compilation starts off from the bare-bones hyphenated, visible-morph-borders lexical transducer - some borders are retained for Hunspell, others for PLX
* based on this, all M4 processing is now removed from the twolc files for sme (sma, smj and others will follow soon)
* to avoid ambiguities (and prepare for future prefixing languages), literal < and > is now encoded as the multicharacter symbol [<] and [>] in the interface between lexc through twolc until they are converted to their true nature at the same point as the inflectional border markers < and > are zeroed out. Without this, all numbers got optional analyses prefixed with both < and > due to the ambiguity between the border markers and the literal strings.

I have tested the transducer before and after the change on a smallish text (which helped me identify a bug), and nothing has changed in the analysed output, except for two new lexicalised compounds showing up. Since the old transducer unfortunately was made last week (and not exactly before I started to make these changes) I don't know whether this is because of new lexicalisations or because of old compounds that was unanalysed earlier because of some strange #-symbol bug. But Thomas has been doing a lot of lexicalisations lately, so I tend to believe it is new lexicalisations. The change is anyway a good one.

I also got some assistance from Tomi on this work.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@54783 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 020777d48ca4c91dd291d5e71ea076ca1e9e14f5