Earlier I made a rule to ensure "Sámedikkiin" was part of the NP (it got X@) and called it RitvaTestComV.
"Gulahallan Sámedikkiin šaddá nu konkrehtalažžan." 

This sentence has come back to haunt me. 

Although it is okay when written in this order (sent-proc.sh -s=dis -t "sentencehere"), the rule is not sufficient enough when we switch the order of the sentence subject 'Gulahallan sámedikkiin': 

"Sámedikkiin gulahallan šaddá nu konkrehtalažžan." 

Here 'Sámediggi' got the @X and an 'Org Build N Pl Loc' reading. We want it to be an adverbial of an infinite verb outside the predicate @-FADVL> with the reading 'N Prop Org Sg Com'.  

Made sure "Sámediggi" is N Prop if "Sámedigge" is written (r853). 

Added a comment to the SELECT:r855 to make it more understandable for myself/others. 

Added comments to SELECT:PropOrNot for the same reason. 

Altered MAP:RitvaTestComV to RitvaTestComVLeft and Right; so ADVL is chosen in both sentences presented. 

Made a REMOVE:RitvaComNotLoc rule to ensure comitative reading of 'sámedikkiin'.