Result of monday syntax meeting, all changes more or less related to vuovdi work:
Added verbs to HUMAN-ACTIVITY-V.

Added two sets, ADJ-LOCATION and NUMAG-ADJ.
We need more such sets, the idea is that different
nouns take different types of adjectives.

Shortened the S-BOUNDARY, S-BOUNDARY1, S-BOUNDARY2
set definitions. The definitions are the same as before,
the difference is in the definitino technique.

New rules for vuovdi, using the sets just defined.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@65297 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 61c782fb5103064185c3f553d59aab93118613c9