Result of monday syntax meeting, all changes more or less related to vuovdi work:
Added verbs to HUMAN-ACTIVITY-V.

Added two sets, ADJ-LOCATION and NUMAG-ADJ.
We need more such sets, the idea is that different
nouns take different types of adjectives.

Shortened the S-BOUNDARY, S-BOUNDARY1, S-BOUNDARY2
set definitions. The definitions are the same as before,
the difference is in the definitino technique.

New rules for vuovdi, using the sets just defined.