The very first step in bringing sme over to the new infrastructure: populating the sme/ directory tree. No source files are added (or if they are added, they are coming from the template, and should be replaced).
NB! The source files *must* be ***svn-copied*** from the old dirs to the new ones - do NOT use regular bash 'cp'. This first version is also only intended for testing - the old infra will remain the production environment for the time being, until all things are working as needed in the new infra.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@70484 c7155fb1-f0a7-4240-a2fc-2600b6f42f90