Wrapped more targets in WANT_PHONETICS, to avoid compilation for people not interested in phonetics.
Added Xerox builds of the phonetic fst's, mainly to help with debugging. It makes it possible to compare the output of both Xerox and Hfst, and check that they are identical (they should of course be). Also - the Xerox build code is much easier to read than the Hfst counterpart, and that might also help in certain cases.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@81705 c7155fb1-f0a7-4240-a2fc-2600b6f42f90