Eliminating MORE flags made the speller FASTER and SMALLER:-D. The speller fst is now one megabyte smaller, and the check&suggestion speed on the typos list is up at 5.18 words/sec! A huge improvement since yesterday evening, when it processed only 1.88 words/sec.
NB! Eliminating ALL flags is counterproductive, and does not yeld (all) these benefits. Specifically, I tried to eliminate the Cap flag (the one controlling capitalisation of derived propers), and that almost doubled the size of the acceptor fst to 67 Mb. It might be that it would have been faster still, but the file size is too big in most cases. That is, this type of fine tuning must be done on a careful per flag approach, testing each flag elimination individually for effects on file size and speller speed.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@123471 c7155fb1-f0a7-4240-a2fc-2600b6f42f90