Test results with the very latest speller, built with eliminate flags on ten different flags. The speed made a big jump up, now at 5.18 w/s, as opposed to 3.29 w/s in the previous version, and 1.88 w/s yesterday evening.
What is very surprising is that the correctness of the language model actually went down, at least according to the typos list - 45 more input words are now accepted compared to the previous version. That should not have been possible, and must be investigated. Also the suggestion quality went down slightly, from 80.90 to 80.77 % in top five. This also needs further investigation.