Left tags are working! That is, when the second part of a compound has a +CmpN/SgGenLeft tag (or one of the other Left variants), that will force the first part of the compound to use the specified case for the compound, independently of the lexically specified or default compounding case for the first part. One step closer to PLX linguistic parity! :D
Turned off flag eliminiation for CmpN flags, due to a bug in Hfst. The bug does not have severe consequences for the language coverage, but I need a stable and predictable version of the speller to work with during development. Keeping the flag means that the speller is very slow ATM, though. When all things are working, the flag will be eliminated again, and the speed should hopefully be back to normal.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@123654 c7155fb1-f0a7-4240-a2fc-2600b6f42f90