New test results with the latest changes to the error model AND a formally correct handling of all compounding restrictions: standard, Left and Def tags are now all handled correctly. In addition, the speed of the speller has been partially restored using flag elimination, done in Foma to avoid bugs in Hfst related to flag semantics during elimination. The result is a speller that is half the speed as the previous (fast but erroneous) speller (compared to a tenth of the speed for some interim versions), and twice the size as well. That is:
speed: 2.52 words/sec (was: 5.63)
suggestion quality: 79.45 % (was: 80.90)

The three correct words not flagged anymore is due to: two words removed from the typos list because they are correct, and a third word that turned out to be correct as well, but not flagged earlier.

Without the help of Foma, we can get a speller that is equally correct and about half the size, but also half the speed.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@124234 c7155fb1-f0a7-4240-a2fc-2600b6f42f90