Added missing +Der tag in front of +Der/XXX tag. You always need one of +Der, +Der1..5 just before the visible, actual derivation tag. This tag is used to both filter out ungrammatical derivations (+Der1..5) and to anchor general operations on derivations without having to list all derivational tags.
Please replace +Der with any of +Der1..5 as suitable. +Der (without any number) means «unspecified for position», ie can appear everywhere. The lexc code may nevertheless restrict the scope of the derivation, such that it in practice is locked in one place. Still, it might be good to add a position tag.