Solved the size issue with the grammar checker analyser - we didn’t block tag renaming across the ‘fake’ compounding for handling sequences of non-compounding nominals to retain that as an alternative analysis for possible compounds written with a space between. That forced the fst to explode in size because it had to consider all possible strange combinations across this border in the compounding lexicon.
I added a tag that is found on this path as a new blocker (in addition to +Cmp for the regular blocker), and the file size for the tokenising analyser went down from 166 Mb to 45 Mb! Quite a nice reduction in size, and I also expect the RAM usage during compilation went down quite a bit (didn’t check).

Now we have a grammar checker analyser & tokeniser that is quite usable size-wise, and very close to being usable linguistically as well (just one issue remains, reported in e-mail to relevant people). A nice way to start the weekend :)