Replaced the old ‘escape’ strings ^excl and ^pipe with their proper characters. It did not cause any issues in the regular preprocess | usme | lookup2cg pipe. There might be other reasons why it has been like this, but I tend to believe that this is just a leftover from older times.
At the same time, the old code made the exclamation mark (and presumably the pipe) disappear from tokenisation output in hfst, which is of course unacceptable. With this commit, tokenisation works as it should.

All in all, no regressions found with my short testing, and one Hfst-based tokenisation bug fixed.

PLEASE TEST!

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@136443 c7155fb1-f0a7-4240-a2fc-2600b6f42f90