New speller test results using the latest speller with the much larger frequency corpus. It did NOT improve the results, although there are several variables:
* the number of typos has increased by 2600 (now 12,5k typos)
* there has been a number of changes to the lexicon and thus the speller

Changes to the speller performance:
* correct sugg in top five has dropped from 79,34 to 78,94
* speed has dropped a bit (not visible in the statistics due to an error, but calculated independently to now be 0.66 words/second vs earlier 0.77)
* some spelling errors accepted as correct - this can also be errors in the typos data
* there is a slight IMPROVEMENT to the average position (from 1.58 to 1.54)

As can be seen from the new statistics, the performance is bad in another area as well: we get the correct suggestion in the first position only in half the cases (53,4% to be exact). This info is not available for the old speller, but the combination of average position and top5 performance indicates that it was actually even worse earlier.