Moved the last file worth keeping to the new infra, deleted gt/sme/src/. Getting closer to emptying gt/sme/:)
The file moved is a shell script to extract all lemmas for use in a wordfeud context, but extracting and listing all lemmas is usable for more purposes. The script needs to be generalised and rewritten, for the time just stored in langs/sme/src/scripts/ until we find a better place and can do the generalisation work.