Committed an experimental build step to build a cg-compatible speller file, se-cg.zhfst. It is NOT automatically built, even when spellers are enabled. To build, cd into tools/spellcheckers/fstbased/desktop/hfst/, and type `make se-cg.zhfst`.
To test the speller with disambiguation of the speller suggestions, run the following command:

$ echo "eallámušvallji eanan šaddá go gehppesmolláneaddji báktešlájat ábaida kálkavallji šlájat váikkuhuvvojit ja nu háddjanit" | preprocess | hfst-ospell-cg se-cg.zhfst | vislcg3 -g disambiguator.cg3

(adjust paths as needed.)

The tool hfst-ospell-cg can be found here: https://github.com/unhammer/hfst-ospell-cg

This is all very experimental, but fun to play with. Eventually something like this will be built into the grammar checker framework, to be used as a more context-intelligent speller. The main idea is to filter the suggestions based on the (relative) immediate context.

NB! The present disambiguator file is pretty effective, and removes too much as it is now. When / if this experiment is turned into a real thing, it will need a dedicated cg file, or something based on a trigram model, or something else that will only remove things we are relatively sure we can throw away. Developing that will be part of the work ahead.