Fixed a number of lexc syntax errors. The errors where revealed by the following error message:
  HINTRSCT generator-raw-gt-desc.tmp1.hfst
/usr/local/bin/hfst-compose-intersect: warning: 
Found output multi-char symbols (">m") in 
transducer in file <stdin> which are not found on the
input tapes of transducers in file phonology/sme-phon.rev.hfst.

It took some binary searching to get to the relevant commit, but then it was easy to see. The error was committed April 6th, and noone had noticed until now…