experiments in tagging space errors
hfst-pmatch seems to handle tagging '( ', but hfst-tokenise treats it
as unmatched even when using the exact same pmhfst, why?

If we fix that, we could simply add explicit analyses of the error
strings in the pmscript like:

  Define spaceafterparenTagged [ "( ":["( " %+SpaceAfterParen ] ];

But we would need to change hfst-tokenise to handle input spanning
lines (currently it only tries to match one line at a time), to handle
the "\n\n" case.