experiments in tagging space errors
hfst-pmatch seems to handle tagging '( ', but hfst-tokenise treats it
as unmatched even when using the exact same pmhfst, why?

If we fix that, we could simply add explicit analyses of the error
strings in the pmscript like:

  Define spaceafterparenTagged [ "( ":["( " %+SpaceAfterParen ] ];

But we would need to change hfst-tokenise to handle input spanning
lines (currently it only tries to match one line at a time), to handle
the "\n\n" case.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@160078 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 68c0e0168b67996c7575046b9874b652c58a739e