dego lávvomuorran: use relations DELETE1 and tag &LINK
Since vislcg3 can't (yet?) do reading-to-reading relations, we need to
help it a bit when there are overlapping suggestions.

To delete a word, in general it's enough to do

  ADD (&some-error) KeepThisWord;
  ADDRELATION (DELETE1) KeepThisWord TO (-1 DeleteThisWord);

but when KeepThisWord has several suggestions on it, e.g. we've also
done

  ADD (&other-error) KeepThisWord;

then we need to additionally do

  ADD (&some-error &LINK) DeleteThisWord IF (1 KeepThisWord);

to make sure we treat the &some-error suggestion as a single error,
separate from the alternative &other-error suggestion.

The &some-error tag on the deleted word is necessary to match up the
readings, since vislg3 relations only go from cohort to cohort, not
reading to reading.

The &LINK on the deleted word is necessary to ensure that KeepThisWord
is the "central" word (with respect to e.g. message templates in
errors.xml), and so we don't get two separate suggestions, but one
merged one.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@160987 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 0edf3308c1a64f29d7dfa9ffe738a65d34b91528