Try to fix generating typos with Ex/A
Revert revision 168940, "had to fix the typo-rule because forms are not generated since tags are missing because of the form of the rule. Things still don't work, but thecause must be in the morphological analyzer."

Not sure what that fixed, but it seems like at least for jierbmát, the
original version I wrote also removed Ex-tags (why?).

After this change, we at least get:

INPUT:
"<jierbmát>"
	"jierbmái" Ex/A Sem/Hum Der/AAdv Adv Err/Orth <W:0.0000000000> <ctjHead> <ctjHead> @ADVL>

OUTPUT:
"<jierbmát>"
        "jierbmái" Ex/A Sem/Hum Der/AAdv Adv Err/Orth <W:0.0000000000> <ctjHead> <ctjHead> @ADVL> &typo ADD:3824:Err/Orth-any
        "jierbmái" Sem/Hum Der/AAdv Adv <W:0.0000000000> <ctjHead> <ctjHead> @ADVL> &typo &SUGGEST ADD:3824:Err/Orth-any COPY:13063:Err/Orth
        "jierbmái" Ex/A Sem/Hum Der/AAdv Adv <W:0.0000000000> <ctjHead> <ctjHead> @ADVL> &typo &SUGGEST ADD:3824:Err/Orth-any COPY:3825:Err/Orth-any

(note there are two groups of Err/Orth-handling rules in
grammarchecker.cg3 – do we want both?)