Added rule as suggested in http://giellatekno.uit.no/bugzilla/show_bug.cgi?id=2466. It works:
$ echo 'Guovdageainnus ja Diehtosiiddas leat maiddai eará guovddáš sámi ásahusat, nu go Sámedikki oahpahus-, giella ja kulturossodat, Sámi árkiiva, Gáldu-Álgoálbmotvuoigatvuođaid gelbbolašvuođaguovddáš ja Riikkaidgaskasaš boazodoalloguovddáš.' | modes/smegram.mode
...
"<giella>"
	"giella" N Sem/Lang_Tool Sg Nom <W:0.0000000000> @COMP-CS< #16->16
: 
"<ja>"
	"ja" CC <W:0.0000000000> @CNP #17->17
: 
"<kulturossodat>"
	"kulturossodat" v1 N Sem/Org Sg Nom <W:0.0000000000> <cohort-with-dynamic-compound> @COMP-CS< #18->18
"<,>"
	"," CLB <W:0.0000000000> #19->19
...

The exact name of the introduced tag is certainly open for discussion, but at least now we have something to base the discussion around.

NB! Note that because the rule is added to mwe-dis.cg3, it is applied both to regular disambiguations (ie corpus analysis) and to the grammar checker analysis. This might be a reason we want to split, or separate out things that are specific to either in a separate file.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@172783 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 107a060b439de09242f5f87f39ca8eb000cbda91