Is it true that the misspelled and the correct form have different transitivities?
Duommá - can you have a look at this? Possible hint: contlex. DOHPPE_IV does not exist, only DOHPPE_TV.

Consequence as it is now: it is not possible to generate the correct form, since the transitivity tag is carried over to the input for the generator. Another solution would of course be to remove the transititivy tags from the generator string, but that could have other side effects in other cases (generating multiple forms when we actually want to generate only one), so it would be better to properly solve the use of the transitivity tag in the misspelled word.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@176842 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: b70f46083554e066fab997f36013c746b4b71ea2