Sorted the entries, kept LIA entries at the bottom. Moved tag lexicons to the top.
Interjections can be used to test the functionality of the pmatch function UpCase, as there are no flag diacritics in them (there is a bug in the present version that blocks UpCase and similar functions if the path contains flag diacritics; the relevan person in the hfst team is aware). Testing so far has shown that UpCase works as it should, except for words containing the sami letters:

Ŋ, Š.

There are no interjections with Đ, Ŧ and Ž, so these are not yet tested.

That is, it works for: Á and Č.

We have discussed adding proper Unicode support to hfst-pmatch, which would solve the issue with only partial support for the Sámi (and other) letters.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@177632 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: bc81246dc89b10a13143a67d1b99de8d4b25062a