Don't add &no-space-after-punct-mark &LINK to every word of the sentence …
a rule like

ADD (tag1 tag2) IF (-1 (tag1));

will keep adding "tag1 tag2" to every word because the previous word had tag1.

We actually just want ot add &no-space-after-punct-mark &LINK if the
previous word was the error origin, ie. didn't have the &LINK tag. So
change the rule to be of the form

ADD (tag1 tag2) IF (-1 (tag1) LINK NOT 0 (tag2));

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@184561 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 2d6865fb6b0ff494e2776063c7ef8f6915f64fa5