Add simple filter to block spellrelax if combined with the Err/Orth-a-á lexicalised spellrelax. Without this one would risk double application, since Err/Orth-a-á is used to tag both changes from a->á, and from á->a. Thus we got changes like this:
a -> á (Err/Orth-a-á), followed by: á -> a (Err/Spellrelax)

That is, perfectly correct forms were double tagged as errors!

Now, using this filter + a slightly adjusted spellrelax-with-flagtags.regex, we avoid this, and only get one or the other.

Spellrelax is still not (re)enabled in the grammar checker, evaluating how much problems it might cause vs potential benefits.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/sme@185342 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 1a36c710f65de58466633d244aea25ae1fc424b9