Speller test results with standard settings, pre-festschrift
Much better than the festschrift experiments with the standard speller, but very slow.

Also much better than the version from two years ago, and compares quite well with the actual festschrift version using the new regex system. That festschrift speller is better on suggestions in first position and on noise, this version is better than the festschrift version in non-first position, and re input with no or only bad suggestions.

The actual festschrift version is als much faster than this speller.